refactor(capabilities): :recycle: migrate effects, form, links, menu, terminal-progress-bar, typography, video molecules to component-store pattern
Effects:
- matrix-background: pure presentational
- matrix-header-background: pure presentational
- matrix-terminal: StateStore<{containerRef,completedLines,currentLine,currentText}> via useMatrixTerminalStore; useGlassmorphism allowlisted in component

Form (all pure presentational):
- form-error-summary, form-field, form-success-message, form-textarea
- form-textarea now uses Label atom for consistency (removes inline label duplication)

Links:
- pills-links (RedPillLink/BluePillLink): pure presentational, multi-export cohesive family kept in one folder
- terminal-link: pure presentational

Menu:
- close: pure presentational
- dropdown-menu: ComponentStore<{open,selected,shouldReduceMotions,buttonRef},{toggleOpen,handleBlur,handleItemClick}> via useDropdownMenuStore; all inline lambdas curried
- hamburger-menu: pure presentational
- menu-item-with-tracking: EffectsStore<{handleClick}> via useMenuItemWithTrackingStore; handleClick curried as (trackingData,callback?)=>()=>void

Remaining:
- terminal-progress-bar: pure presentational; already folder-composition compliant; index.ts added
- page-title, paragraph-title-with-icon: pure presentational
- self-hosted-video, youtube: pure presentational
