+++ src/components/design-system/molecules/effects/matrix-background/index.ts
@@ -1,0 +1,1 @@
+export { MatrixBackground } from "./matrix-background";
