+++ src/components/design-system/molecules/effects/matrix-header-background/index.ts
@@ -1,0 +1,1 @@
+export { MatrixHeaderBackground } from "./matrix-header-background";
