+++ src/components/design-system/molecules/effects/matrix-header-background/matrix-header-background.tsx
@@ -1,0 +1,22 @@
+import { FC } from "react";
+import { MatrixRain } from "@/components/design-system/atoms/effects/matrix-rain/matrix-rain";
+
+interface MatrixHeaderBackgroundProps {
+    big: boolean;
+}
+
+export const MatrixHeaderBackground: FC<MatrixHeaderBackgroundProps> = ({ big }) => {
+    const height = big
+        ? "h-[350px] sm:h-[400px] md:h-[500px]"
+        : "h-[240px] sm:h-[250px] md:h-[270px]";
+
+    return (
+        <div
+            className={`absolute -z-10 ${height} top-0 left-0 right-0 overflow-hidden border-b-2 border-solid border-accent shadow-lg`}
+        >
+            <div className="absolute top-0 left-0 right-0 bottom-0 remove-scroll-width">
+                <MatrixRain />
+            </div>
+        </div>
+    );
+};
