+++ src/components/design-system/molecules/effects/matrix-terminal/index.ts
@@ -1,0 +1,1 @@
+export { MatrixTerminal } from "./matrix-terminal";
