+++ src/components/design-system/molecules/form/form-error-summary/index.ts
@@ -1,0 +1,2 @@
+export { FormErrorSummary } from "./form-error-summary";
+export type { FormErrorSummaryProps } from "./form-error-summary";
