+++ src/components/design-system/molecules/form/form-field/index.ts
@@ -1,0 +1,2 @@
+export { FormField } from "./form-field";
+export type { FormFieldProps } from "./form-field";
