+++ src/components/design-system/molecules/form/form-success-message/index.ts
@@ -1,0 +1,2 @@
+export { FormSuccessMessage } from "./form-success-message";
+export type { FormSuccessMessageProps } from "./form-success-message";
