+++ src/components/design-system/molecules/form/form-textarea/index.ts
@@ -1,0 +1,2 @@
+export { FormTextarea } from "./form-textarea";
+export type { FormTextareaProps } from "./form-textarea";
