+++ src/components/design-system/molecules/links/pills-links/index.ts
@@ -1,0 +1,1 @@
+export { RedPillLink, BluePillLink } from "./pills-links";
