+++ src/components/design-system/molecules/links/pills-links/pills-links.tsx
@@ -1,0 +1,23 @@
+"use client";
+
+import { TrackingData } from "@/types/configuration/tracking";
+import { FC, PropsWithChildren } from "react";
+import { BluePill, RedPill } from "@/components/design-system/atoms/effects/pills";
+import { StandardInternalLinkWithTracking } from "@/components/design-system/atoms/links/standard-internal-link-with-tracking";
+
+type PillProps = PropsWithChildren<{
+    to: string;
+    trackingData: TrackingData;
+}>;
+
+export const RedPillLink: FC<PillProps> = ({ children, to, trackingData }) => (
+    <StandardInternalLinkWithTracking className="no-underline" to={to} trackingData={trackingData}>
+        <RedPill>{children}</RedPill>
+    </StandardInternalLinkWithTracking>
+);
+
+export const BluePillLink: FC<PillProps> = ({ children, to, trackingData }) => (
+    <StandardInternalLinkWithTracking className="no-underline" to={to} trackingData={trackingData}>
+        <BluePill>{children}</BluePill>
+    </StandardInternalLinkWithTracking>
+);
