+++ src/components/design-system/molecules/links/terminal-link/index.ts
@@ -1,0 +1,1 @@
+export { TerminalLink } from "./terminal-link";
