+++ src/components/design-system/molecules/menu/close/index.ts
@@ -1,0 +1,1 @@
+export { Close } from "./close";
