+++ src/components/design-system/molecules/menu/dropdown-menu/dropdown-menu.tsx
@@ -1,9 +1,11 @@
+"use client";
+
 import { TrackingData } from "@/types/configuration/tracking";
 import { AnimatePresence } from "framer-motion";
-import { FC, useEffect, useRef, useState } from "react";
+import { FC } from "react";
 import { BiChevronDown } from "react-icons/bi";
-import { useReducedMotions } from "@/components/design-system/hooks/use-reduced-motions";
-import { MenuItemWithTracking } from "./menu-item-with-tracking";
+import { MenuItemWithTracking } from "@/components/design-system/molecules/menu/menu-item-with-tracking";
+import { useDropdownMenuStore } from "./use-dropdown-menu-store";
 
 interface DropdownMenuItem {
     label: string;
@@ -36,34 +38,21 @@ export const DropdownMenu: FC<DropdownMenuProps> = ({
     className = "",
     chevronClassName = "",
 }) => {
-    const shouldReduceMotions = useReducedMotions();
-    const [open, setOpen] = useState(false);
-    const buttonRef = useRef<HTMLButtonElement>(null);
-    const selected = items.flatMap((entry) =>
-        isDropdownMenuGroup(entry) ? entry.items : [entry],
-    ).some((item) => item.selected);
-
-    useEffect(() => {
-        if (!open) return;
-        const handleScroll = () => setOpen(false);
-        window.addEventListener("scroll", handleScroll, { passive: true });
-        return () => window.removeEventListener("scroll", handleScroll);
-    }, [open]);
-
-    const handleBlur = (e: React.FocusEvent<HTMLButtonElement | HTMLDivElement>) => {
-        if (!e.currentTarget.contains(e.relatedTarget as Node)) {
-            setOpen(false);
-        }
-    };
+    const hasSelected = items
+        .flatMap((entry) => (isDropdownMenuGroup(entry) ? entry.items : [entry]))
+        .some((item) => item.selected);
+    const { state, effects } = useDropdownMenuStore(hasSelected);
+    const { open, selected, shouldReduceMotions, buttonRef } = state;
+    const { toggleOpen, handleBlur, handleItemClick } = effects;
 
     return (
-        <div className={`relative z-50 mb-0`} tabIndex={-1} onBlur={handleBlur}>
+        <div className="relative z-50 mb-0" tabIndex={-1} onBlur={handleBlur}>
             <button
                 ref={buttonRef}
                 className={`${className} xs:pl-4 xs:pr-1 xs:py-1 hover:bg-accent-alpha-10 hover:text-accent hover:border-accent relative flex flex-nowrap items-center justify-center gap-2 rounded-xl border border-solid px-1 py-2 text-center text-sm leading-normal text-shadow-md md:text-base ${open || selected ? "border-accent bg-accent-alpha-15 text-accent" : "border-transparent"}`}
                 aria-haspopup="menu"
                 aria-expanded={open}
-                onClick={() => setOpen((v) => !v)}
+                onClick={toggleOpen}
                 type="button"
             >
                 <span>{label}</span>
@@ -75,7 +64,7 @@ export const DropdownMenu: FC<DropdownMenuProps> = ({
             <AnimatePresence>
                 {open && (
                     <div
-                        key={"dropdown-menu"}
+                        key="dropdown-menu"
                         className={`glow-container ${shouldReduceMotions ? "xs:bg-general-background" : "xs:bg-general-background/90"} relative mt-2 w-auto min-w-max rounded-xl py-2 xs:absolute xs:right-0 xs:left-0 xs:w-auto`}
                         tabIndex={-1}
                         role="menu"
@@ -96,9 +85,7 @@ export const DropdownMenu: FC<DropdownMenuProps> = ({
                                             trackingData={item.trackingData}
                                             selected={item.selected ?? false}
                                             className="xs:whitespace-nowrap m-2 text-center"
-                                            onClickCallback={() => {
-                                                item.onClickCallback?.();
-                                            }}
+                                            onClickCallback={handleItemClick(item.onClickCallback)}
                                         >
                                             {item.label}
                                         </MenuItemWithTracking>
@@ -111,13 +98,11 @@ export const DropdownMenu: FC<DropdownMenuProps> = ({
                                     trackingData={entry.trackingData}
                                     selected={entry.selected ?? false}
                                     className="xs:whitespace-nowrap m-2"
-                                    onClickCallback={() => {
-                                        entry.onClickCallback?.();
-                                    }}
+                                    onClickCallback={handleItemClick(entry.onClickCallback)}
                                 >
                                     {entry.label}
                                 </MenuItemWithTracking>
-                            ),
+                            )
                         )}
                     </div>
                 )}
