+++ src/components/design-system/molecules/menu/dropdown-menu/index.ts
@@ -1,0 +1,1 @@
+export { DropdownMenu } from "./dropdown-menu";
