+++ src/components/design-system/molecules/menu/hamburger-menu/hamburger-menu.tsx
@@ -2,9 +2,9 @@ import { BiMenu } from "react-icons/bi";
 import { FC } from "react";
 
 interface HamburgerMenuProps {
-  onClick: () => void;
+    onClick: () => void;
 }
 
 export const HamburgerMenu: FC<HamburgerMenuProps> = ({ onClick }) => (
-  <BiMenu className="size-9" onClick={onClick} />
+    <BiMenu className="size-9" onClick={onClick} />
 );
