+++ src/components/design-system/molecules/menu/hamburger-menu/index.ts
@@ -1,0 +1,1 @@
+export { HamburgerMenu } from "./hamburger-menu";
