+++ src/components/design-system/molecules/menu/menu-item-with-tracking/index.ts
@@ -1,0 +1,1 @@
+export { MenuItemWithTracking } from "./menu-item-with-tracking";
