+++ src/components/design-system/molecules/terminal-progress-bar/index.ts
@@ -1,0 +1,1 @@
+export { TerminalProgressBar } from "./terminal-progress-bar";
