+++ src/components/design-system/molecules/typography/page-title/index.ts
@@ -1,0 +1,1 @@
+export { PageTitle } from "./page-title";
