+++ src/components/design-system/molecules/typography/page-title/page-title.tsx
@@ -1,5 +1,5 @@
 import { FC, PropsWithChildren } from "react";
 
 export const PageTitle: FC<PropsWithChildren> = ({ children }) => (
-  <h1 className="mb-4">{children}</h1>
+    <h1 className="mb-4">{children}</h1>
 );
