+++ src/components/design-system/molecules/typography/paragraph-title-with-icon/index.ts
@@ -1,0 +1,1 @@
+export { ParagraphTitleWithIcon } from "./paragraph-title-with-icon";
