+++ src/components/design-system/molecules/typography/paragraph-title-with-icon/paragraph-title-with-icon.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+export const ParagraphTitleWithIcon: React.FC<{
+    icon: React.ReactNode;
+    children: string;
+}> = ({ icon, children }) => (
+    <span className="flex items-center gap-2">
+        {icon} {children}
+    </span>
+);
