+++ src/components/design-system/molecules/video/self-hosted-video/index.ts
@@ -1,0 +1,1 @@
+export { SelfHostedVideo } from "./self-hosted-video";
