--- src/components/design-system/molecules/video/youtube.tsx
@@ -1,8 +1,0 @@
-import { FC } from "react";
-
-interface Props {
-  videoId: string;
-}
-
-export const Youtube: FC<Props> = ({ videoId }) => 
-  <iframe src={`https://www.youtube.com/embed/${videoId}`} title="React Native multiple debugger" loading="lazy" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerPolicy="strict-origin-when-cross-origin" allowFullScreen></iframe>
