+++ src/components/design-system/molecules/video/youtube/index.ts
@@ -1,0 +1,1 @@
+export { Youtube } from "./youtube";
